rename SwapInfo to FrameInfo
As discussed on the mailing list we can be slightly more consistent with gtk
by using "frame" instead of "swap" here. Also we discussed the potential
for being able to query information for a frame that's still in progress
and in that case "swap" wouldn't make much sense.

TODO: squash this back into owen's patch