Add cogl_frame_info_get_output()
If available from the windowing system, it could potentially be useful
to know something other than the refresh rate of the output that the
monitor is currently presenting on. (Note however, an onscreen may span
multiple outputs. The windowing system has to pick *one* for timing,
but most other attibutes of the window will apply to the parts of the
onscreen on that output.)