Don't use g_get_monotonic_time()
Avoid a hard dependency on GLib and on however GLib is implementing
g_get_monotonic_time() by directly referencing clock_gettime(CLOCK_MONOTONIC).

The translation of timestamp values for unidentified types of UST is
also removed, since it's going to have a big error, and
glXGetSyncValues() doesn't return the *current* UST anyways - it returns
the UST for the last frame.