gl: consider min glsl version for core profile
We can't assume that drivers have support for glsl 1.2 when we are
requesting to use the core profile for GL 3 drivers. This makes cogl
request the minimum glsl version that we we know must be available for
core profile drivers.

This is required to use Cogl on later versions of OSX.