vulkan: add cogl-vulkan-symbols.x X macro file
The file lists the Vulkan symbols we need to look up in libvulkan.so at
runtime in the form of SYM(vkSymbolName) declarations. The file can
be included for different definitions of SYM().

This is similar to what we do for GL API to help minimize the amount of
boilerplate we need to write manually for each symbol.